// src/types.ts
export type Currency = 'AFN' | 'USD' | 'EUR' | 'PKR';

export interface RoznamchaEntry {
  id: number;
  date: string;
  type: 'income' | 'expense';
  amount: number;
  description: string;
  bill_number?: string;
  customer_id?: number;
  currency?: Currency;
}
// ...
export interface Stats {
  totalIncome: number;
  totalExpense: number;
  totalStockIn: number;
  totalStockOut: number;
  balance: number;
}